// software/src/ui/pages/layout-editor-page/models/LayoutManagerBase.ts
import { ILayoutEditSource } from '~/shared';

export type ILayoutManagerEditTarget = 'CurrentProfile' | 'LayoutFile';

export type ILayoutManagerModalState =
  | 'None'
  | 'LoadFromProject'
  | 'SaveToProject';
export const layoutManagerState = new (class {
  layoutEditSource: ILayoutEditSource = { type: 'CurrentProfile' };
  modalState: ILayoutManagerModalState = 'None';
})();